Add OR to items, clean up changes
Adds OR search to the items page. Probably not all that useful, but needed to make the table compatible with updated search filter anyway.

Reworded tips on both the crew and items table, and tried to use examples that demonstrate handling of unusual search terms, like T'Pol (apostrophe) and Q Continuum (multiword trait).

Renamed the matches variables to meetsAnyCondition and meetsAllConditions for clarification. Changed OR regex to handle spaces better.