List optimal polestar combinations
This identifies the polestar combinations that give the best chance at getting a specified crew from crew retrieval. Right now it favors combinations with the fewest polestars as possible, though for some crew it's still a pretty big list.

Note that this isn't very pretty and there are likely more efficient or more useful ways to do this. We could possibly simplify the constellation section on the crew page and combine the optimal polestars with that. Any advice on improving this are welcome!