Eliminate supersets when precalculating perfect retrievals
This should eliminate redundant polestar combinations that are supersets of other perfect subsets, when precalculating each crew's perfect combinations.

It's basically the same logic as what's on the individual crew page, only this time using an inline function. I couldn't think of any way to make this more efficient, but I'm open to ideas!