New Player Tools landing page, with stored sessions
Created a new player tools landing page (playertools.tsx). It's functionally the same as the previous (voyage.tsx), but reworks a lot of the internals to match the structure of the profile page and to make it easier to maintain going forward. All components have been moved to the landing page, with a new component created for the unneeded items tool (unneededitems.tsx) and a new component that's exclusively dedicated to the voyage calculator (voyagecalculator_iap.tsx). The top menu now points to the new player tools page and the address will be datacore.app/playertools. (The existing page for tools and the component for the voyage calculator remain in place until the new landing are fully reviewed, after which we leave them up for legacy or redirect to the new landing page.)

All player tools now work with the same customized playerData variable (preparedProfileData) and future components should use that instead of  any of the version of playerData for consistency.

A stripped version of the playerData (the same format that is used when sharing profiles) is now saved in session storage, so that a user will not have to re-import the player.json file if they leave the player tools. The time of the import is now shown on the landing page, with a button to show the upload form again if they want to update. This player data only resides in session until the user closes the browser.

To do: copy the export button from the profile page over to the landing page, and maybe copy an 'Update now' button from the landing page to the profile page.

Voyage Calculator fixes:
The voyage calculator now properly considers frozen crew when that option is toggled.
The excludes form now only shows crew that are on your active roster by default. When the consider frozen crew option is checked, the excludes shows all crew accordingly.
Also reworded the options to say 'consider' instead of 'include', for clarity.

Top menu change: Removed content from popup, as that was triggering a react warning.