Voyage calculator fixes
Fixed 2 bugs with the voyage calculator that I missed, both related to `preparedProfileData` not having the same info as `playerData`.

1) The voyage calculator should now properly consider `active` crew that are on shuttles or voyages, when that option is toggled. Had to change playertools.tsx accordingly to pass along the relevant data from playerData. Renamed the `peopleListStyle` state from `active` to `default` to distinguish it from crew who are actively on shuttles or voyages.

2) The voyage calculator should now properly consider crew proficiencies when recommending and giving an estimate time. Recommendations should now perfectly match the ones given from the existing voyage calculator component.

Other notable changes: The player tools import timestamp is now a dropdown button with the option to "Update now" and "Clear player data." Not sure if I like this better or not than the single update button.

And some minor changes: `PLAYERLINK` is now a global constant in playertools.tsx in anticipation of adding other device instructions. The `lastImported` property now better matches the `lastModified` property generated for profiles. Renamed a session variable to `strippedPlayerData` for consistency.

In unneededitems.tsx, deleted some commented code that was meant for testing.