Fix ship level display in player utils
This has been a long-standing display glitch, which has shown the levels of owned ships always by 1 lower than they should be.

To clarify: all ship data from the `player.json` game export as well as the `ship_schematics.json` start counting levels at 0 (to a maximum of 9 for a 5* ship). This has resulted in ships being shown as level 9 / 9 when they are maxed, but more importantly owned 1 / 10 and un-owned 0 / 10 ships have both been erroneously displayed as 0 / 9.