Simplifies constellation details, adds icons, alternates to optimal polestars list
On the crew page, this simplifies the constellation details into a grid of polestar icons rather than a table with polestar icons and flavors.

Optimal polestars now use icons instead of just text. Additionally if a crew has a non-100% chance of being retrieved, the chance also lists the alternate crew members that might be retrieved with the same polestar combination.

Optimal polestars now use the precalculated list of unique combos if it exists. Otherwise it will generate the best chance combos on demand.

Adds a small description to optimal polestars: "The best chance to retrieve this crew using as few polestars as possible"