Player Tools refactor - new files only
This is #74, but split up into 2 parts, all based on a more recent master. This first half isolates new files from changed files. Until the second half is merged, the refactored player tools can only be accessed directly at /playertools

Integrated new changes since the latest commit to #74:
#104 Added Player tools title to layout property for SEO
#116 Added new dialog with abort button to voyage calculator component

Topline changes from #74:
Created a new player tools landing page (playertools.tsx). It's functionally similar to the previous one, but reworks a lot of the internals to match the structure of the existing profile page and to make it easier to maintain going forward. All components have been moved inside the landing page, with a new component created for the unneeded items tool (unneededitems.tsx) and a new component that's exclusively dedicated to the voyage calculator (voyagecalculator_iap.tsx).

All player tools now work with the same customized playerData variation (generated by prepareProfileData). Future components should use that instead of any other versions of player data whenever possible.

A stripped version of player data (the same format that is used when sharing profiles) is now saved in session storage, so that a user will not have to re-import the player.json file if they click away from the player tools. A timestamp of the import is now shown on the landing page and acts as a button to allow users to update or clear their player data. This player data only resides in session until the user closes the browser.

This adds a new tier to Unneeded Items so that unneeded items are presented as 1) unneeded schematics, 2) unneeded crew-specific items, and 3) unneeded generic items that could be useful later. It also makes the tool a little more conservative when considering items.

Fixes #15, #21, #46, #73