Player Tools refactor - updated links
This updates all appropriate links to point to the refactored Player Tools' new address at `/playertools`, specifically links in:
* `topmenu.tsx`
* `fleet_info.tsx`
* `profile.tsx`

If a user visits the old address at `/voyage`, they are presented with a simple notice to update their bookmarks to `/playertools`.

This deletes the `voyagecalculator.tsx` file; as of #120, all tools components are now called from the `playertools.tsx` page, and the actual calculator logic is now in `voyagecalculator_iap.tsx`.