New experimental voyage calculator
Here's a new experimental voyage calculator, written in plain JS and using a very different algorithm. The good news is that in most of my tests so far, the results are on par or better than the current "classic" calculator and they're delivered reliably quickly. The bad news is that the new spinner doesn't spin quite so good because plain, single-threaded JS blocks the UI as the calculator is doing its thing.

I put together a new player tools page that allows you to run each calculator side by side and compare the results. This will also reassess any classic results with the runtime estimate you'd get by feeding the numbers into the bot, which is more pessimistic than the website's classic estimate.

This is nowhere near ready for production and requires a lot more testing, specifically lots and lots of player data. Please compare as many voyages as you can and report any bugs, inconsistencies, or times where the experimental recommendations are worse than the classic.

Thanks!