Estimation handled by worker
The voyage runtime estimator has been moved to a new web worker so that the calculation is done in the background and avoids locking the UI. The worker can calculate a single estimation (which it does now when we reassess a classic calculation). It can also determine the best estimation from multiple lineups (which is preferred by the experimental calculator).

Not sure if this is the best way to structure or instantiate the worker. Also need to track down a reference error that's triggered when the worker is called. Any advice on this would be great.