Adds search box to profile crew page
This adds a search box to the crew roster component on player tools and profiles, but there's a lot going on under the hood in this one:

This refactors searchabletable and profile_crew to use React hooks. Among other benefits, this will allow us to more easily tie state data to session, thus letting us store form data on any page that needs it. You can see it in action right now with the refactored profile crew page.

A new storage utils file holds the logic for the storage hook. It uses sessionStorage by default, but it should work with localStorage too. Hopefully this will be fully compatible with other components, but some may need to be refactored first.

Moved some logic to searchabletable and the new crewsearch utils file, so that searchabletable can handle sorting and filtering by search string for more components (in this case, profile_crew but hopefully crew retrieval down the line). The main index and items page were updated to be compatible with these changes.

The profile crew component now has a checkbox to toggle frozen crew.

But really this is all just to see if I can create a new Cloudflare test page from a PR.