Various fixes
Profile_crew: removed code that added a dummy field to player data in favor of using the default skill properties for sorting; this should fix the sorting problems that the dummy field created in profile_crew2 (which was triggered by visiting the crew mobile tool after the crew tool, an uncaught error created in #164).

Profile_crew2: this adds a quick fix to #3. The page renders before the list of equipment image URLs has been fetched, which is why the images are broken. This simply shows the default equipment placeholder image while the fetch is occurring, and then properly updates the images once the list of URLs is ready. There are better ways to approach this, but this should work for now. This should also fix an issue with some high-level equipment not showing.