Standardizes session variable names
Separated player tools panes so that certain stored variables are only set/called when player data is found. This lets us add DBID to variable names, to distinguish settings when there are multiple profiles on the same device.

Also attempts to standardize stored variable names in the format {dbid}/{pageName}/{settingName} for longterm storage, or {pageName}/{settingName} for session storage.