Event planner, shuttle helper player tools
Adds a new "Event planner" player tool that shows a searchable crew table with options specific to event planning (e.g. show event crew for an upcoming event, apply event bonus to skill scores). The Event planner tool also includes an experimental "Shuttle helper" that allows users to test shuttle setups for faction events.

Player tools now checks for the playerData array glitch on import (see #207) and treats the first element of array as the actual playerData object. Everything seems to work as expected from this point, but we should keep an eye out for issues here because of our limited ability to test this properly.

After pasting, the clipped content more explicitly notes that DataCore is intentionally displaying less data in the textbox to speed up the process.

Moved the import form out of a React container. Not sure if there's any reason why it should be in one, we're not using containers for any other pages (except BB). There's extra padding on the container, so removing the container here lets the elements after the import form line up as expected.

Tweaked storage so that sessionStorage values are initialized more immediately and that localStorage values (which must be async) are initialized with an optional callback message. This should allow us to better control renders (i.e. now we can prevent the page from showing a "wrong" setting value while we're waiting for the preferred value to be loaded from storage). It also checks for window now to avoid test-build errors.