Skirmish fix, bonus column, consider active, remember setups
Player tools does not appear to identify event crew for skirmishes until the event is live, so had to add a check for that to make sure event planner tool still loads properly, even if no event crew are found.

Crew table:
Added Bonus column to show each crew's bonus multiplier. Note that the multiplier is always faction-based (x2, x3) rather than event specific. The table should be sorted by the bonus column by default, but need to change searchabletable first to accept to default sort options.

Shuttle helper:
Added option to consider (or not consider) active crew. Had to change playertools to keep active crew as a separate variable in session; should change voyagecalculator to use the activeCrew variable for consistency.

Added option to remember shuttle setups between sessions. This is the first instance of a toggable remember option: my solution here is to make the setting always use localStorage, but to clear it on session close if necessary.