Fixes for CAB rank, behold roster links
CAB rank is by rarity, so this adds a missing filter to the link.

"On your roster" ranks have a different crew pool and they should link to the playertool/crew page instead of the main crew table. Because there are no dedicated player tool URLs yet, this disables roster rank links on the behold page for the time being.