Updated baseTargets
Updated MVAM to use new baseTargets calculated from simulations of Joshurtree's revised Chewable. Rewrote the lineup generator to scale attempts to roster size and to determine the initial origin from roster strength with vectors being a small range of primeFactors (instead of being random origins and a mostly fixed primeFactor). Boosts are now determined by primeFactor deltas. Proximity and deviation calculations moved to the estimator.

Rewrote the MVAM estimator class with a new default setting ‘auto’ for proximityThreshold. When set to auto, the estimator radically narrows the number of lineups to estimate based on the average proximity and scores that are within 1 standard deviation of the average score. The highest option for proximityThreshold has been set to 1.

Added a `postscript` property to the calculator helper that allows for notes to be added to the bottom of the results pane. MVAM now uses postscript to explain why it returned multiple results.

Fixed a bug so that voyage telemetry data is properly assembled and sent. Hopefully.