Adds announcement message to index
This is just a quick mockup of how we can handle announcement messages on DataCore's front page.

Announcements are written as markdown files in the static folder with fields for customizing the header, semantic icon, and semantic class. Right now only the most recent announcement will appear and it's present for only 3 days beyond the posted date.

Some potential areas for improvement:
* adding a template so we can link an excerpted message to a page with full details (or to a page that shows all messages).
* adding the ability to dismiss announcements