Various sort fixes
Fixes sorting by rarity on crew retrieval tool.

Fixes CAB sorting on all searchabletables so that crew without a CAB rating are moved to the back of a descending sort (and shown first on an ascending sort), which makes the ordering more predictable, i.e. treating unrated crew as if their CAB rating was 0.

Fixes big book tier sorting on all searchabletables so that crew without a tier rating are moved to the back of BOTH sorting directions. This is very hacky, but preferable because the big book's real value is on an ascending sort. Otherwise unrated crew tiers are -1 (or sometimes undefined), which would give them priority over Tier 1 crew on an ascending sort.

Replaced underscores on pseudocolumn names as shown on searchabletables.