Reorganized About menu
Turns `About` link in the menu into a submenu with link to new announcements page and consolidated "All other pages" items

Moves up Behold and Player tools in the menu.

Simplified item generation a little to make it easier to add links in the future. Uses link-based keys instead of index for better reliability.