Assorted fixes to MVA calculator
Changes to MVA lineup algorithm:
Sets boundaries on boosts to avoid swingy-ness that led to bad lineup recommendations (or outright crashing) in certain cases; also lowers finetune threshold to 0.05. This should result in slightly more thorough lineup generation.

Changes to MVA analyzer algorithm:
The analyzer no longer assumes that "good" lineups were generated. This should prevent the calculator from crashing in that case, though this may result in some bad recommendations (a problem that would ideally be fixed in the lineup generator).

Because there were changes made to the generator algorithm, I made the analyzer a little more conservative. As a result, some recommendations will take a little longer. We can optimize this again later.

UI changes:
The recommendation descriptions now show the values (e.g. time) alongside the recommendation.

Reverted crew popup to show skill numbers again when hovering over crew names in a recommended lineup. Not sure what the case is for `remapSkills` in `voyageutils`; I left the underlying code there just in case, but is it needed?

Added the word `crew` to a button to distinguish it from the in voyage calculator.

Added a `<Header>` to playertools just above each component. With the tab-navigation removed, the tools felt like they were missing a page identifier; we also needed a block-level element to properly shift the elements down and obey expected margins, which the header will do for us.