Additional optimizations
Based on feedback that some people aren't getting good recommendations from MVA, this increases the number of lineups generated by bumping the maxAttempts up to 25% of roster size (from 20%) and minUniques up to 25% of maxAttempts (from 20%). This will increase the odds that a better lineup will be considered. It should not noticeably change the total time of calculation.

Tweaks best ship recommendation to use the ship with the fewest taps, if multiple ships would otherwise be considered equally "best."