Fix voyage calc antimatter count
Fixes starting antimatter count on original voyage calculator by using `traits` (symbols) instead of `traits_named` to match traits, which should handle multi-word traits (and renamed traits) more consistently.

Fixes seat order on voyage editor by sorting to match `CONFIG` crew slots (in cases when `voyageConfig` order is different than expected order).

Re-orders `SKILLS` constants in `CONFIG` to match in-game order. Here, this will show the aggregate skill numbers in the same order as the in-game voyage screen. This does not affect the order of skills as shown in searchabletables, which should be changed individually.

Based on feedback that recommendation descriptions still aren't as clear as they ought to be, this changes some terms to be more self-explanatory ("estimated runtime" instead of "median runtime", "guaranteed minimum" instead of "guaranteed runtime").