Fix potential skills of FF crew in event planner
Fix event planner to properly calculate potential skills of fully-fused but unleveled crew.

Uses a Gatsby <Link> to link crew names in the table to crew pages. This allows users to middle-click the names to open up pages in a new tab, similar to some changes made in #315.

Skills headings in tableConfig are now dynamically generated so that they match skill order as defined in CONFIG. This means we can re-order the skills in CONFIG in the future without needing to update eventplanner individually.