Skill depth table added to crew player tool
New potential featured added to crew player tool: skill depth table. This table shows the depth and strength of your roster at various areas of the game for every relevant skill combination. With it you can more easily identify your weak spots, especially when compared to your own roster's strengths ("Best Ten") as well as the best crew in the game ("Percentile").

Uses a Gatsby `<Link>` to link crew names in the table to crew pages. This allows users to middle-click the names to open up pages in a new tab, similar to changes made in #315 and #317. Main crew stats page (index) is next.

Skills headings in tableConfig are now dynamically generated so that they match skill order as defined in CONFIG. This means we can now re-order the skills in CONFIG without needing to update child components individually, similar to changes made in #317. Main crew stats page (index) is next.

Updated searchableTable to accept search parameters in the URL (and state). Probably not all that useful for profilecrew but this is also prepwork for broader use with main crew stats page.