Limit fuse combos in crew retrieval (#315)
This limits the number of combo options to 200 per fuse group; it also limits calculation of options for up to 5 fuses. This should speed up viewing combos of certain crew members that have lots and lots of possible options without hurting usefulness that much. If slowdowns continue to be a problem or if we really want to display every single option, we might consider paging the results. This functionality is also a good candidate for a worker.

Uses a Gatsby <Link> to link crew names in the table to crew pages instead of using the onClick handler. This allows users to middle-click the names to open up pages in a new tab. We should update all searchable tables (e.g. index, crew player tool, event planner) to act the same way.