Prospective crew added to crew tool
This adds the ability to add prospective crew to the main crew player tool. First introduced in the event planner, this feature is useful for seeing how new crew might fit into your existing roster. The skill depth table (previously introduced in this branch) will also change to reflect prospective crew.

A new `prospectpicker` component was created to handle prospective crew. TODO: Update the event planner to use the new component.

The `searchabletable` component has a new feature: lockable crew. This adds a row of buttons to the table, so that you can quickly locate specific crew (here, prospective crew) in the table. The table will automatically paginate to the page with the locked crew when you change filter or sort options. This makes it easier to directly compare prospective crew to the crew on your roster that are most similar to the prospect.

TODO: Update `commoncrewdata` to use lockable crew, which is more robust than the previous `highlights` implementation

TODO: Also bring over the option to show potential skills of unleveled crew from the event planner.

Renamed some session variables for consistency.

Removes the word "legacy" from the tier statistic. Should Tier be re-added as a column in the table?