Catch new keystones on crew page
With this, the full crew page should now fully render even if the keystone list is outdated (e.g. after a server update adds a new keystone but before DataCore properly regenerates the keystones list). This guesses the keystone name and icon to use in the crew's optimal polestar table from the polestar symbol.

This fix also indexes a few mapped elements to remove the logged error about missing keys when rendering.