Add preview button to behold tool
Adds a button on the behold page to let you preview the choices on your roster (i.e. add them as prospects to the crew player tool). A similar button can be added to individual crew pages.

Changed the prospect picker so that the list is only populated on demand. This should speed up the initial rendering of any page that uses the prospect picker.

Cleaned up some logic.