Big Book notes/crew table changes
Removes Big Book legacy tags and discontinued notes from index, crew page, crew player tool, crew retrieval, behold tool, and about page. Adds relevant links to bigbook.app from crew and behold pages, as requested. Re-adds Tier column to crew table on index page.

Uses a Gatsby `<Link>` to link crew names in the crew tables (on index and crew player tool) to their respective crew pages. This allows users to middle-click the names to open up pages in a new tab, similar to changes already approved in #315 and #317.

Skills headings in `tableConfig` (on index and crew player tool) are now dynamically generated so that they match skill order as defined in CONFIG. This means we can now re-order the skills in CONFIG without needing to update child components individually, similar to changes approved in #317.

Updates searchableTable to accept search parameters from the URL and state. Changes highlighting logic and introduces "lockable" crew buttons to the top of the table to quickly jump to highlighted crew.

Adds the ability to add prospective crew to the crew player tool (using a new prospectpicker component).

Adds a button on the behold page to let you preview the choices on your roster (i.e. adds them as prospects to the crew player tool by passing params to searchabletable in state).

Some of these changes were first demoed in #322, but I moved them here because they affected many of the same files. The skill depth table introduced in #322 is NOT included here as it still needs work and will probably be moved to its own page.