Custom columns, permalinks for crew tables
Added a couple small things to this that were on the todo list.

Allows for custom columns on the main crew page table. Right now, this adds a column when you click a rank on a crew page that's not already in the table (e.g. gauntlet rank). This is also prep work for enabling advanced options on crew tables in the future.

Allows for permalinks to be added to the bottom of a searchabletable, so that users can bookmark or share the current view of the table (i.e. it creates an URL that includes current search terms, sorting, and paging options).

Adds the Tier column and Collections pseudocolumn to the crew table player tool, to match the layout of the index crew table.

Adds a preview in roster button to the crew template page.

Cleaned up some parts and consolidated reusable code into the searchabletable component.