Checks for unneeded higher-level items of a similar name
Adds another filter to the fodder check that assumes an item is needed if a higher quality item of the same name is needed. For example, 3* Kira's Pistol (which is unneeded on its own for any crew on my roster) will no longer be listed because it's similar in name to 4* Kira's Pistol (which is needed on my roster by an unequipped Decoy Kira).

This will lead to some false positives (e.g. some 2* items are not actually building blocks of their equivalent 3* builds), but that's okay because it's better to be too conservative in recommendations than the alternative. Plus, the false positives will resolve themselves when the relevant crew become fully equipped.

Unfortunately, the tool still cannot account for partial builds by different names, e.g. 3* Tuvix's Uniform needs 2* Starfleet Security Uniform (VOY), in which case the latter might still be listed as fodder even if Tuvix is unequipped. This is also the case first raised in #285 about Yar's Phaser being a building block of Armus' Skin. To reduce the chances of this, lower quality equipment (i.e. 0* and 1*) remain ignored by the final tier of recommendations (Other Equipment). The new disclaimer is hopefully enough to warn users of this possibility.

If DataCore's items.json template gets updated with an incremental recipe tree, the `needsHigherQuality` function can be replaced with one that properly checks higher-level recipes for needed building blocks.

Made small changes to some variable names and the disclaimer for clarity.