DataLore: a Wordle clone for STT
No one asked for it, but here it is anyway: a Wordle clone for Star Trek Timelines! Guess the mystery crew in 8 tries or less, using your knowledge of the crew's series, rarity, skills, and traits to help narrow the possibilities. There's a daily version that challenges everyone to the same mystery crew each day, and a practice version that you can play as often as you like. Access the game from the Pages menu > DataLore.

This is a proof of concept, with rules and gameplay subject to change. This also may not be a proper fit for DataCore. I just thought it would be fun to try. Inspired by Wordle, of course, but more specifically the baseball-themed WARdle.