Adds close button, hardcode fixes
Adds a dedicated close button to the crew picker modal and defaults the primary action button to close the modal when no crew is selected.

Hardcodes fixes for known crew with the wrong series. Ultimately decided that this was the best way to handle this. There are only 20 instances of crew with the wrong series (or missing a series). The fixes for these were all unambiguous and in line with the unofficial Trait Audit thread.

Makes another pass at handling variants, opting to use the same logic as the crew page. (So Anastasia is once again a proper Kira variant, etc.) But this also means including a hardcode exception for dax variants (filtering by short_name), otherwise all Daxes would be variants of one another. One last variant thing that could be improved in the future: the crew picker doesn't handle variants at all (e.g. Anastasia won't appear if you search for "Kira").

Reduces the likelihood of repeat daily solutions by checking against likely solutions from the previous 3 weeks.