Various fixes
Handles 0 events in the player data more gracefully.

Correctly identifies active status of crew from player data.

Adds a brief explanation of the lock function to the assignments description. Shows a lock icon for locked assignments in the assignment picker (note the lock here does not preclude a user from reassigning that crew from this modal).

Adds an option to 'Select only open in-game' to the checkmark menu of the missions list (which is to say, ignore missions previously imported by the shuttle helper).