Share Worfle results, customize rules
Adds a button to let players copy results to clipboard for sharing.

Adds the ability to customize rules for practice games, letting players change the max guesses and filter the pool by series and rarities.

Adds a note to the crew picker to make it clear that only crew in the time portal are valid guesses.

There will be 2 mystery crew on days when the time portal is updated. The eligible pool are all crew with the `in_portal` flag set to true at the time the game is initiated by the player. A player who initiates a daily game after the time portal is updated will get a different mystery crew than a player who initiates it beforehand. Made changes here to better handle the days when the portal changes (i.e. if you start before the change, your solution remains the previous solution rather than the new solution).

Rather than use `in_portal,` we can at some point switch to using `date_added` to filter the pool, setting it to, say, exclude any crew added within the last month. This should eliminate the double mystery crew problem described above because this would be less prone to the random timing of portal updates. But this might make repeat solutions more likely because it's more complicated to determine the solutions from the past month.