Fix estimate of ongoing voyage
This should fix the estimate of an ongoing voyage in the voyage calculator. The corrected duration is now properly limited to voyages with a detected duration bug.

Couple of other fixes here: this now properly sends the voyage score (rather than the aggregate object) to Chewable when doing the estimate. This will affect the estimate. Also cleans up the message about a detected duration bug.