A few fixes for trait filtering
The trait options in the filter by trait dropdown are now sorted properly. Users can also now type in the dropdown to search for traits.

Matching traits are now shown in the description label beneath crew names.

Added a level pseudocolumn to the crew column so that the table can be sorted by crew level.