Trait buttons +?, reset button, zero message
The Combo Wizard modal shows the buttons in a grid for a cleaner look. The buttons now say Trait + ? + ? to show how many mystery traits are needed for that node, rather than a number.

Adds a message to all searchable tables when 0 results are found. The zero message can be customized by individual searchable tables as needed.

Adds a Reset button to the end of the filter options row. The button appears when fields are autopopulated by the FBB wizard (and the zero message also suggests resetting after using a wizard).

Clarifies some filter options to 'Show all *owned* crew.'