Unowned table changes, element generalizations
Changes the behavior of when the unowned crew table is shown automatically: 1) always when using the combo wizard, 2) when there are unowned crew but not matching crew on roster, and 3) when the user opens it manually. The user now always has control over showing/hiding the unowned crew table.

Adds a dropdown to set minimum number of trait matches. This is only shown when there are at least 2 requested trait filters.

Moves a few elements into generalized component pages so they can be reused more easily by other tools, e.g. the index page when it gets updated to allow for ship abilities and filtering by traits.

The fleet boss player data is now taken from the whole fbb root rather than the statuses branch. This makes it possible for future tools to access valor quantity and the boss name.

Fixes crewsearch so that the index page can properly search by ship abilities.

Adds style fix to ensure that the filter row is not partially hidden by the combo wizard button on mobile.