Announcements UI changes
Changes the announcements message to look a little better, especially in dark mode.

Uses the class name to color the message rather than use the built-in semantic slate style, which tends to be overwhelming in dark mode. Supported classes here are: info, warning, positive, negative. No class is also acceptable.

Increases the size of the message text and turns the details link into a button. Both cases increase readability.