Shows rarity of matched traits
Adds colors to the trait match labels to indicate trait "rarity" among matches (e.g. legendary = only crew that matched that trait, super rare = 1 of 2 crew who matched trait, etc). This is not the most efficient way to do this, but we're limited right now by how searchabletables work.

Sets minimum to number of matched traits to match the number of needed traits.