Sets dates for 2* Lower Decks crew
This manually sets the dates added for Ensigns Boimler, Mariner, Rutherford, and Tendi, introduced in the Lower Decks mega. The dates will take effect when precalculate is next run.

See #373 for earlier notes on dates and 2* crew. Will have to do again in a couple of weeks for new 2* Kevin Mulkahey.