Mark crew as solution
Adds ability to mark crew as a solution. A modal will appear when clarification is needed. (Note: this does not handle crew when they are the solution to multiple nodes simultaneously.)

Crew checklist moved to below crew table, where it more logically fits.

Superset is ignored if it only belongs to crew not in portal.

Crew table should now remember previous settings when data is change. On the other hand, table is no longer sorted initially by coverage.

Hidden traits now always appear as ? in table, even if partially solved. If/when partially-solved nodes are fully supported, this can be reverted.

Prepwork for exporting data (traits should be functional but not yet displayed).

Moved subcomponents to their own file for easier maintenance. Various optimizations.